linkdine_scrapper/pyproject.toml
[project]
name = "deep-learning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.121.2",
    "groq>=0.33.0",
    "langchain>=1.0.3",
    "langchain-openai>=1.0.2",
    "matplotlib>=3.10.7",
    "openai>=2.7.1",
    "playwright>=1.56.0",
    "playwright-stealth>=2.0.0",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
]


